Move to r2dbc 1.0.0

+++ src/test/java/dev/miku/r2dbc/mysql/IntegrationTestSupport.java
@@ -63,7 +63,7 @@ abstract class IntegrationTestSupport {
             .as(StepVerifier::create);
     }
 
-    static Mono<Integer> extractRowsUpdated(Result result) {
+    static Mono<Long> extractRowsUpdated(Result result) {
         return Mono.from(result.getRowsUpdated());
     }
